third_party/blink/web_tests/platform/generic/touchadjustment/editable-content-expected.txt

Hello

World

Tests touch adjustment on a text-area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test textarea
test direct touches
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
test indirect touches
PASS adjusted node was TEXTAREA#text-area.
PASS adjusted node was TEXTAREA#text-area.
PASS adjusted node was TEXTAREA#text-area.
PASS adjusted node was TEXTAREA#text-area.
test richly editable div
test direct touches
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 0
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 1
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
PASS Math.floor((adjustedPoint.y - origin) / rowHeight) is 2
test indirect touches
PASS adjusted node was DIV#rich-text.
PASS adjusted node was DIV#rich-text.
PASS adjusted node was DIV#rich-text.
PASS adjusted node was DIV#rich-text.
PASS successfullyParsed is true

TEST COMPLETE
